Convert SearchColumnDropdown to typescript
